refactor: Refactored links into own component
