+++ src/data/links.ts
@@ -1,0 +1,13 @@
+  const links = [
+    { href: '/', label: 'Home' },
+    { href: '/#about', label: 'About' },
+    { href: '/#faqs', label: 'FAQs' },
+    { href: '/#sponsors', label: 'Sponsors' },
+    { href: '/past-winners', label: 'Past Winners' },
+    { href: '/theme-history', label: 'Theme History' },
+    { href: '/gallery', label: 'Gallery' },
+    { href: '/organisers', label: 'Organisers' },
+    { href: '/registration', label: 'Registration' },
+  ];
+
+  export default links
